fix(discussions): thread blocks the whole UI (modal Radix Sheet overlay) — render non-modally in the ActivityPanel
